+++ mapreader/annotate/annotator.py
@@ -512,7 +512,12 @@ class Annotator:
         queue_df["eligible"] = queue_df.apply(check_eligibility, axis=1)
 
         if self._sortby is not None:
-            queue_df.sort_values(self._sortby, ascending=self._ascending, inplace=True)
+            queue_df.sort_values(
+                by=[self._sortby, "min_y"],
+                ascending=[self._ascending, True],
+                kind="mergesort",
+                inplace=True,
+            )
             queue_df = queue_df[queue_df.eligible]
         else:
             queue_df = queue_df[queue_df.eligible].sample(frac=1)  # shuffle