+++ CHANGELOG.md
@@ -14,9 +14,16 @@ The following table shows which versions of MapReader are compatible with which 
 ## Pre-release
 _Add new changes here_
 
+### Fixed
+
+- Fixes the `model_summary` method in the `ClassifierContainer` class ([#574](https://github.com/maps-as-data/MapReader/pull/574))
+
+## [v1.8.2](https://github.com/Living-with-machines/MapReader/releases/tag/v1.8.2) (2025-12-19)
+
 ### Added
 
 - Added Hugging Face model support to ClassifierContainer
+- Added `piffle` package as dependency ([#575](https://github.com/maps-as-data/MapReader/pull/575))
 
 ## [v1.8.1](https://github.com/Living-with-machines/MapReader/releases/tag/v1.8.1) (2025-08-11)